requirements.txt
numpy==1.26.0
opencv-python==4.8.1.78
PySDL2==0.9.16
torch==2.1.0  
torchvision==0.16.0